Add GetContinuations as next building block for path search alg.
Regardless of whether we implement BFS, Dijkstra or A*,
we will need a way to find the next possible path
directions from a given cell.

This helper will find the correct (output connected) neighbor
for a given cell, and return all 5 legal cell variants for
that location, or none if the target location is blocked or out of bounds.